% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Variable Labels from the ABC-DS Data Dictionary}
\usage{
apply_labels(data, study, codebook)
}
\arguments{
\item{data}{A dataset existing in the user's global environment for which
to apply variable labels.}

\item{study}{A symbol or unquoted name of the study on the ATRI EDC from which
to request the codebook.}

\item{codebook}{A symbol or unquoted name of the codebook within the specified
study to request.}
}
\value{
A data frame identical to the input dataset, but with variable-level
\code{label} attributes applied based on the corresponding entries in the
ABC-DS data dictionary.
}
\description{
Applies human-readable variable labels to an existing dataset using the
ABC-DS (Alzheimer’s Biomarker Consortium–Down Syndrome) data dictionary
retrieved from the ATRI EDC system.
}
\details{
This function uses \code{\link{get_codebook}} to retrieve the specified study’s
data dictionary from the ATRI EDC and applies the descriptive field labels
to variables in the dataset.
}
\examples{
\dontrun{
# Example: Apply variable labels to an existing dataset
demographics <- apply_labels(demographics, abcds, ptdemog)
}

}
\seealso{
\code{\link{get_codebook}},
\code{\link{apply_factor_labels}}
}
