% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{atri_get}
\alias{atri_get}
\title{Make a GET Request to the ATRI EDC API}
\usage{
atri_get(server, token)
}
\arguments{
\item{server}{A character string containing the fully formatted server URL
pointing to the folder or file location on the ATRI EDC data lake.}

\item{token}{A character string containing the user's API token for
authenticating with the ATRI EDC API.}
}
\value{
An \code{httr2} response object.
}
\description{
Sends a GET request to the ATRI EDC API using a formatted server URL and a
user token, returning the API response.
}
\details{
This function constructs and performs a GET request to the ATRI EDC API using
the provided server URL and authentication token. It is intended for internal
use by higher-level functions that retrieve study data.
}
\seealso{
\code{\link[httr2]{request}},
\code{\link[httr2]{req_headers}},
\code{\link[httr2]{req_perform}}
}
\keyword{internal}
