% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{FUNCTION_TITLE}
\usage{
generate_report(..., outputdir, lonidir = NULL)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{outputdir}{PARAM_DESCRIPTION}

\item{lonidir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cli]{cli_abort}}
\code{\link[glue]{glue}}
\code{\link[yaml]{write_yaml}}
\code{\link[quarto]{quarto_render}}
}
