% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_atri_files}
\alias{get_atri_files}
\title{Retrieve ATRI EDC File Links}
\usage{
get_atri_files(study, ...)
}
\arguments{
\item{study}{The name of the study on the ATRI EDC for which to request files.}

\item{...}{Additional subfolders or path segments appended to the server path
to locate specific files.}
}
\value{
A character vector containing API links to the folders or files stored in the ATRI EDC.
}
\description{
Retrieves API links for folders and files stored in the ATRI EDC for a given study,
optionally navigating through additional subfolders.
}
\details{
The \code{get_atri_files()} function constructs API links by combining the base
server path for a study (retrieved via \code{\link{get_atri_server}}) with any
additional folder or file path components provided via \code{...}. These links
can then be used to retrieve data from the ATRI EDC using \code{\link{import_atri_file}}.
}
\seealso{
\code{\link[rlang]{ensym}},
\code{\link[rlang]{is_symbol}},
\code{\link[rlang]{as_string}},
\code{\link{import_atri_file}},
\code{\link{get_atri_server}}
}
