% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_atri_token}
\alias{get_atri_token}
\title{Retrieve ATRI API Token}
\usage{
get_atri_token(study)
}
\arguments{
\item{study}{The name of the study on the ATRI EDC for which the token is
requested.}
}
\value{
A character string containing the formatted API token suitable for making
calls to the ATRI EDC data lake.
}
\description{
Retrieves the API token for the current user from the R environment for a
specified ATRI EDC study.
}
\details{
This function looks for the API token in the user's R environment and formats
it for use in ATRI EDC API calls. It is intended for internal use by other
functions in the package that interact with ATRI EDC data.
}
\keyword{internal}
