% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{FUNCTION_TITLE}
\usage{
get_consensus(..., site = NULL, cycle = NULL, apply_labels = FALSE)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{site}{PARAM_DESCRIPTION, Default: NULL}

\item{cycle}{PARAM_DESCRIPTION, Default: NULL}

\item{apply_labels}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{defusing-advanced}}
\code{\link[tidyr]{pivot_wider}}
\code{\link[dplyr]{reexports}}
}
