% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_demographics}
\alias{get_demographics}
\title{Retrieve and Format ABC-DS Demographic Data}
\usage{
get_demographics(
  ...,
  site = NULL,
  cycle = NULL,
  apply_labels = FALSE,
  controls = FALSE
)
}
\arguments{
\item{...}{One or more unquoted variable names to select from the demographics
dataset.}

\item{site}{An optional site code or name used to filter the demographic data;
defaults to \code{NULL}, returning all sites.}

\item{cycle}{An optional study cycle identifier used to filter the data;
defaults to \code{NULL}, returning all cycles.}

\item{apply_labels}{Logical. Whether to apply variable and factor labels from
the ABC-DS data dictionary. Defaults to \code{FALSE}.}

\item{controls}{A boolean value that indicates whether the function should return the controls, Default: FALSE}
}
\value{
A \code{\link[tibble]{tibble}} containing selected demographic variables,
optionally filtered by site or study cycle. If \code{apply_labels = TRUE},
the variables will include descriptive labels and factor levels.
}
\description{
Retrieves participant demographic data from the ABC-DS study within the ATRI
Electronic Data Capture (EDC) system and returns a clean, labeled dataset
optionally filtered by site or study cycle.
}
\details{
The \code{get_demographics()} function simplifies extraction of participant-level
demographic data for ABC-DS analyses. Users can specify particular variables of
interest, optionally limit results by site or cycle, and automatically attach
human-readable labels using \code{\link{apply_labels}} and
\code{\link{apply_factor_labels}}.

If \code{age_at_visit} is among the requested variables, the function will
automatically compute age at visit based on the participant’s year of birth
and evaluation date (\code{de_eval_date}). This ensures accurate age
calculation even when the variable is not explicitly stored in the dataset.

Internally, the function uses \pkg{purrr} for functional mapping and reshapes
the data with \pkg{tidyr} as needed for analysis-ready output.
}
\examples{
\dontrun{
# Retrieve basic demographics for all sites
demo <- get_demographics(age_at_visit, de_race, de_gender, de_ethnicity)

# Retrieve labeled demographics for a specific site and cycle
demo_site <- get_demographics(
  age_at_visit, de_race, de_gender, de_ethnicity,
  site = "KUMC", cycle = 2, apply_labels = TRUE
)
}

}
\seealso{
\code{\link[purrr]{map}},
\code{\link[tibble]{as_tibble}},
\code{\link[tidyr]{pivot_wider}},
\code{\link{apply_labels}},
\code{\link{apply_factor_labels}}
}
