% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure.R
\name{get_disclosure}
\alias{get_disclosure}
\title{Retrieve the Disclosure Variables from ATRI}
\usage{
get_disclosure(site = NULL, cycle = NULL, apply_labels = FALSE)
}
\arguments{
\item{site}{An optional site code or name used to filter the demographic data;
defaults to \code{NULL}, returning all sites.}

\item{cycle}{An optional study cycle identifier used to filter the data;
defaults to \code{NULL}, returning all cycles.}

\item{apply_labels}{Logical. Whether to apply variable and factor labels from
the ABC-DS data dictionary. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[tibble]{tibble}} containing selected disclosure variables,
optionally filtered by site or study cycle. If \code{apply_labels = TRUE},
the variables will include descriptive labels and factor levels.
}
\description{
Retrieves variables needed for the BRIDGE21 Down syndrome
disclosure report from the ATRI Electronic Data Capture system
}
\details{
The \code{get_disclosure()} function simplifies extraction of participant-level
data being disclosed in the BRIDGE21 report. Users can limit results by site or
cycle, and automatically attach human-readable labels using
\code{\link{apply_labels}} and \code{\link{apply_factor_labels}}.
}
\seealso{
\code{\link[purrr]{reduce}}
}
