% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sibling_controls}
\alias{get_sibling_controls}
\title{Retrieve the identifiers for the sibling controls}
\usage{
get_sibling_controls()
}
\value{
A character vector containing a list of sibling identifiers
}
\description{
Retrieve Sibling Identifiers
}
\details{
Retrieve Sibling Identifiers
}
