% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_from_environment}
\alias{retrieve_from_environment}
\title{Retrieve ATRI Token or Server from Environment}
\usage{
retrieve_from_environment(study, type = c("token", "server"))
}
\arguments{
\item{study}{The name of the study on the ATRI EDC for which data is requested.}

\item{type}{The type of environment variable to return. Must be either
\code{"token"} or \code{"server"}. Default: \code{c("token", "server")}.}
}
\value{
A character string containing the formatted name of the environment variable
needed to access the ATRI API.
}
\description{
Attempts to locate the ATRI API token or server in a user's R environment
based on the specified study.
}
\details{
This function searches the user's R environment for the ATRI token or server
associated with a given study. If the variable is not found, it will throw
an informative error using \code{cli_abort}. Useful for programmatic access
to ATRI API resources without hardcoding credentials.
}
\seealso{
\code{\link[cli]{cli_abort}},
\code{\link[glue]{glue}}
}
\keyword{internal}
